Use MUnit in ConnectionImplSuite
