+++ http/src/test/scala/org/http4s/blaze/http/http2/CodecUtils.scala
@@ -21,12 +21,7 @@ import java.nio.charset.StandardCharsets
 
 import org.http4s.blaze.util.BufferTools._
 
-import scala.concurrent.{Await, Awaitable}
-import scala.concurrent.duration._
-
 private[http2] object CodecUtils {
-  def await[T](awaitable: Awaitable[T]): T = Await.result(awaitable, 5.seconds)
-
   def byteData: Array[Byte] =
     "The quick brown fox jumps over the lazy dog".getBytes(StandardCharsets.UTF_8)
